+++ core/src/main/scala/com/github/gvolpe/fs2rabbit/algebra/Acking.scala
@@ -16,9 +16,9 @@
 
 package com.github.gvolpe.fs2rabbit.algebra
 
-import com.github.gvolpe.fs2rabbit.model.StreamAcker
+import com.github.gvolpe.fs2rabbit.model.Acker
 import com.rabbitmq.client.Channel
 
-trait Acker[F[_]] {
-  def createAcker(channel: Channel): F[StreamAcker[F]]
+trait Acking[F[_]] {
+  def createAcker(channel: Channel): F[Acker[F]]
 }
