+++ core/src/main/scala/com/github/gvolpe/fs2rabbit/algebra/Publishing.scala
@@ -25,7 +25,7 @@ trait Publishing[F[_], G[_]] {
       channel: Channel,
       exchangeName: ExchangeName,
       routingKey: RoutingKey
-  ): F[StreamPublisher[G]]
+  ): F[Publisher[G]]
 
   def createPublisherWithListener(
       channel: Channel,
@@ -33,6 +33,6 @@ trait Publishing[F[_], G[_]] {
       routingKey: RoutingKey,
       flags: PublishingFlag,
       listener: PublishingListener[G]
-  ): F[StreamPublisher[G]]
+  ): F[Publisher[G]]
 
 }
