+++ core/src/main/scala/com/github/gvolpe/fs2rabbit/interpreter/Fs2Rabbit.scala
@@ -38,8 +38,8 @@ object Fs2Rabbit {
     ConnectionStream.mkConnectionFactory[F](config, sslContext).map { factory =>
       val amqpClient = new AMQPClientStream[F]
       val connStream = new ConnectionStream[F](factory)
-      val acker      = new AckerProgram[F](config, amqpClient)
-      val consumer   = new ConsumerProgram[F](amqpClient)
+      val acker      = new AckingProgram[F](config, amqpClient)
+      val consumer   = new ConsumingProgram[F](amqpClient)
       new Fs2Rabbit[F](config, connStream, amqpClient, acker, consumer)
     }
 }
@@ -49,12 +49,12 @@ class Fs2Rabbit[F[_]: Concurrent] private[fs2rabbit] (
     config: Fs2RabbitConfig,
     connectionStream: Connection[Stream[F, ?]],
     amqpClient: AMQPClient[Stream[F, ?], F],
-    acker: Acker[F],
-    consumer: Consumer[Stream[F, ?], F]
+    acker: Acking[F],
+    consumer: Consuming[Stream[F, ?], F]
 ) {
 
-  private[fs2rabbit] val consumingProgram: Consuming[Stream[F, ?], F] =
-    new ConsumingProgram[F](acker, consumer)
+  private[fs2rabbit] val consumingProgram: AckConsuming[Stream[F, ?], F] =
+    new AckConsumingProgram[F](acker, consumer)
 
   private[fs2rabbit] val publishingProgram: Publishing[Stream[F, ?], F] =
     new PublishingProgram[F](amqpClient)
@@ -65,7 +65,7 @@ class Fs2Rabbit[F[_]: Concurrent] private[fs2rabbit] (
       queueName: QueueName,
       basicQos: BasicQos = BasicQos(prefetchSize = 0, prefetchCount = 1),
       consumerArgs: Option[ConsumerArgs] = None
-  )(implicit channel: AMQPChannel, decoder: EnvelopeDecoder[F, A]): Stream[F, (StreamAcker[F], StreamConsumer[F, A])] =
+  )(implicit channel: AMQPChannel, decoder: EnvelopeDecoder[F, A]): Stream[F, (Acker[F], StreamConsumer[F, A])] =
     consumingProgram.createAckerConsumer(channel.value, queueName, basicQos, consumerArgs)
 
   def createAutoAckConsumer[A](
@@ -76,7 +76,7 @@ class Fs2Rabbit[F[_]: Concurrent] private[fs2rabbit] (
     consumingProgram.createAutoAckConsumer(channel.value, queueName, basicQos, consumerArgs)
 
   def createPublisher(exchangeName: ExchangeName, routingKey: RoutingKey)(
-      implicit channel: AMQPChannel): Stream[F, StreamPublisher[F]] =
+      implicit channel: AMQPChannel): Stream[F, Publisher[F]] =
     publishingProgram.createPublisher(channel.value, exchangeName, routingKey)
 
   def createPublisherWithListener(
@@ -84,7 +84,7 @@ class Fs2Rabbit[F[_]: Concurrent] private[fs2rabbit] (
       routingKey: RoutingKey,
       flags: PublishingFlag,
       listener: PublishingListener[F]
-  )(implicit channel: AMQPChannel): Stream[F, StreamPublisher[F]] =
+  )(implicit channel: AMQPChannel): Stream[F, Publisher[F]] =
     publishingProgram.createPublisherWithListener(channel.value, exchangeName, routingKey, flags, listener)
 
   def addPublishingListener(listener: PublishingListener[F])(implicit channel: AMQPChannel): Stream[F, Unit] =
