+++ site/src/main/tut/publishers/json.md
@@ -6,7 +6,7 @@ number: 12
 
 # Publishing Json
 
-A stream-based `Json Encoder` that can be connected to a `StreamPublisher` is provided by the extra dependency `fs2-rabbit-circe`. Implicit encoders for your classes must be on scope. You can use `Circe`'s codec auto derivation for example:
+A stream-based `Json Encoder` that can be connected to a `Publisher` is provided by the extra dependency `fs2-rabbit-circe`. Implicit encoders for your classes must be on scope. You can use `Circe`'s codec auto derivation for example:
 
 ```tut:book:silent
 import cats.effect.IO
@@ -20,7 +20,7 @@ case class Person(id: Long, name: String, address: Address)
 
 object ioEncoder extends Fs2JsonEncoder[IO]
 
-def program(publisher: StreamPublisher[IO]) = {
+def program(publisher: Publisher[IO]) = {
   import ioEncoder._
 
   val message = AmqpMessage(Person(1L, "Sherlock", Address(212, "Baker St")), AmqpProperties.empty)
